+++ CHANGELOG.md
@@ -74,6 +74,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Show query state, project bridgehead state, datashield state and coder state only for one bridgehead
 - Creator Name and creator email button
 - User name and email button
+- Hide project rows in dialog if they are not editable except on summary
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)